Fix operation stats for map and replicatedMap
Map.clear and Map.containsKey, ReplicatedMap.get were not updating
the stats.
ReplicatedMap.clear were updating the stats twice.

Map.getAll/Map.putAll stats for client were missing.

Fixed the issues above and added tests for member and client.

fixes https://github.com/hazelcast/hazelcast/issues/14143
