+++ hazelcast-client/src/test/java/com/hazelcast/client/map/ClientMapStatsTest.java
@@ -1,0 +1,62 @@
+/*
+ * Copyright (c) 2008-2018, Hazelcast, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hazelcast.client.map;
+
+import com.hazelcast.client.test.TestHazelcastFactory;
+import com.hazelcast.core.HazelcastInstance;
+import com.hazelcast.core.IMap;
+import com.hazelcast.map.LocalMapStatsTest;
+import com.hazelcast.monitor.LocalMapStats;
+import com.hazelcast.test.HazelcastParallelClassRunner;
+import com.hazelcast.test.annotation.ParallelTest;
+import com.hazelcast.test.annotation.QuickTest;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.experimental.categories.Category;
+import org.junit.runner.RunWith;
+
+
+@RunWith(HazelcastParallelClassRunner.class)
+@Category({QuickTest.class, ParallelTest.class})
+public class ClientMapStatsTest extends LocalMapStatsTest {
+    private TestHazelcastFactory factory = new TestHazelcastFactory();
+
+    private String mapName = "mapName";
+    private HazelcastInstance client;
+    private HazelcastInstance member;
+
+    @Before
+    public void setUp() {
+        member = factory.newHazelcastInstance();
+        client = factory.newHazelcastClient();
+    }
+
+    @After
+    public void cleanup() {
+        factory.terminateAll();
+    }
+
+    @Override
+    protected LocalMapStats geMapStats() {
+        return member.getMap(mapName).getLocalMapStats();
+    }
+
+    @Override
+    protected <K, V> IMap<K, V> getMap() {
+        return client.getMap(mapName);
+    }
+}
