+++ hazelcast/src/test/java/com/hazelcast/map/LocalMapStatsTest.java
@@ -16,45 +16,52 @@
 
 package com.hazelcast.map;
 
-import com.hazelcast.config.Config;
-import com.hazelcast.config.EvictionPolicy;
-import com.hazelcast.config.MapConfig;
-import com.hazelcast.config.MaxSizeConfig;
-import com.hazelcast.config.MemberGroupConfig;
-import com.hazelcast.config.PartitionGroupConfig;
-import com.hazelcast.core.EntryEvent;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.IMap;
-import com.hazelcast.core.MultiMap;
-import com.hazelcast.internal.partition.InternalPartitionService;
-import com.hazelcast.map.listener.EntryEvictedListener;
 import com.hazelcast.monitor.LocalMapStats;
-import com.hazelcast.nio.Address;
-import com.hazelcast.spi.properties.GroupProperty;
 import com.hazelcast.test.AssertTask;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.HazelcastTestSupport;
-import com.hazelcast.test.TestHazelcastInstanceFactory;
 import com.hazelcast.test.annotation.ParallelTest;
 import com.hazelcast.test.annotation.QuickTest;
 import com.hazelcast.util.Clock;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.CountDownLatch;
 
+import static com.hazelcast.spi.properties.GroupProperty.PARTITION_COUNT;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 @RunWith(HazelcastParallelClassRunner.class)
 @Category({QuickTest.class, ParallelTest.class})
 public class LocalMapStatsTest extends HazelcastTestSupport {
+
+    private static final int OPERATION_COUNT = 10;
+    private static final int DEFAULT_PARTITION_COUNT = Integer.valueOf(PARTITION_COUNT.getDefaultValue());
+
+    private HazelcastInstance instance;
+    private String mapName = "mapName";
+
+    @Before
+    public void setUp() {
+        instance = createHazelcastInstance(getConfig());
+    }
+
+    protected LocalMapStats geMapStats() {
+        return instance.getMap(mapName).getLocalMapStats();
+    }
+
+    protected <K, V> IMap<K, V> getMap() {
+        warmUpPartitions(instance);
+        return instance.getMap(mapName);
+    }
 
     @Test
     public void testHitsGenerated() throws Exception {
@@ -63,7 +70,7 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
             map.put(i, i);
             map.get(i);
         }
-        LocalMapStats localMapStats = map.getLocalMapStats();
+        LocalMapStats localMapStats = geMapStats();
         assertEquals(100, localMapStats.getHits());
     }
 
@@ -74,7 +81,7 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
             map.put(i, i);
             map.get(i);
         }
-        LocalMapStats localMapStats = map.getLocalMapStats();
+        LocalMapStats localMapStats = geMapStats();
         assertEquals(100, localMapStats.getPutOperationCount());
         assertEquals(100, localMapStats.getHits());
     }
@@ -85,7 +92,7 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
         for (int i = 0; i < 100; i++) {
             map.putAsync(i, i);
         }
-        final LocalMapStats localMapStats = map.getLocalMapStats();
+        final LocalMapStats localMapStats = geMapStats();
         assertTrueEventually(new AssertTask() {
             @Override
             public void run()
@@ -102,7 +109,7 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
             map.put(i, i);
             map.get(i);
         }
-        LocalMapStats localMapStats = map.getLocalMapStats();
+        LocalMapStats localMapStats = geMapStats();
         assertEquals(100, localMapStats.getGetOperationCount());
         assertEquals(100, localMapStats.getHits());
     }
@@ -116,7 +123,7 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
             putMap.put(100 + i, 100 + i);
             map.putAll(putMap);
         }
-        LocalMapStats localMapStats = map.getLocalMapStats();
+        LocalMapStats localMapStats = geMapStats();
         assertEquals(200, localMapStats.getPutOperationCount());
     }
 
@@ -132,7 +139,7 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
             keys.add(100 + i);
             map.getAll(keys);
         }
-        LocalMapStats localMapStats = map.getLocalMapStats();
+        LocalMapStats localMapStats = geMapStats();
         assertEquals(200, localMapStats.getGetOperationCount());
     }
 
@@ -148,7 +155,7 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
             @Override
             public void run()
                     throws Exception {
-                final LocalMapStats localMapStats = map.getLocalMapStats();
+                final LocalMapStats localMapStats = geMapStats();
                 assertEquals(100, localMapStats.getGetOperationCount());
                 assertEquals(100, localMapStats.getHits());
             }
@@ -162,7 +169,7 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
             map.put(i, i);
             map.remove(i);
         }
-        LocalMapStats localMapStats = map.getLocalMapStats();
+        LocalMapStats localMapStats = geMapStats();
         assertEquals(100, localMapStats.getRemoveOperationCount());
     }
 
@@ -173,7 +180,7 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
             map.put(i, i);
             map.removeAsync(i);
         }
-        final LocalMapStats localMapStats = map.getLocalMapStats();
+        final LocalMapStats localMapStats = geMapStats();
         assertTrueEventually(new AssertTask() {
             @Override
             public void run()
@@ -191,7 +198,7 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
             map.put(i, i);
             map.get(i);
         }
-        final LocalMapStats localMapStats = map.getLocalMapStats();
+        final LocalMapStats localMapStats = geMapStats();
         final long initialHits = localMapStats.getHits();
 
         new Thread(new Runnable() {
@@ -200,7 +207,7 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
                 for (int i = 0; i < actionCount; i++) {
                     map.get(i);
                 }
-                map.getLocalMapStats(); // causes the local stats object to update
+                geMapStats(); // causes the local stats object to update
             }
         }).start();
 
@@ -224,12 +231,12 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
         map.put(key, "value");
         map.get(key);
 
-        long lastAccessTime = map.getLocalMapStats().getLastAccessTime();
+        long lastAccessTime = geMapStats().getLastAccessTime();
         assertTrue(lastAccessTime >= startTime);
 
         Thread.sleep(5);
         map.put(key, "value2");
-        long lastAccessTime2 = map.getLocalMapStats().getLastAccessTime();
+        long lastAccessTime2 = geMapStats().getLastAccessTime();
         assertTrue(lastAccessTime2 > lastAccessTime);
     }
 
@@ -242,7 +249,7 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
         map.put(key, "value");
         map.put(key, "value");
 
-        final LocalMapStats localMapStats = map.getLocalMapStats();
+        final LocalMapStats localMapStats = geMapStats();
         final long lastUpdateTime = localMapStats.getLastUpdateTime();
 
         new Thread(new Runnable() {
@@ -250,7 +257,7 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
             public void run() {
                 sleepAtLeastMillis(1);
                 map.put(key, "value2");
-                map.getLocalMapStats(); // causes the local stats object to update
+                geMapStats(); // causes the local stats object to update
             }
         }).start();
 
@@ -270,171 +277,105 @@ public class LocalMapStatsTest extends HazelcastTestSupport {
         map.put("key", "value");
         map.evictAll();
 
-        final long heapCost = map.getLocalMapStats().getHeapCost();
+        final long heapCost = geMapStats().getHeapCost();
 
         assertEquals(0L, heapCost);
     }
 
+
     @Test
-    public void testHits_whenMultipleNodes() throws InterruptedException {
-        TestHazelcastInstanceFactory factory = createHazelcastInstanceFactory(2);
-        final HazelcastInstance[] instances = factory.newInstances(getConfig());
-        MultiMap<Object, Object> multiMap0 = instances[0].getMultiMap("testHits_whenMultipleNodes");
-        MultiMap<Object, Object> multiMap1 = instances[1].getMultiMap("testHits_whenMultipleNodes");
+    public void testOtherOperationCount_containsKey() {
+        Map map = getMap();
 
-        // InternalPartitionService is used in order to determine owners of these keys that we will use.
-        InternalPartitionService partitionService = getNode(instances[0]).getPartitionService();
-        Address address = partitionService.getPartitionOwner(partitionService.getPartitionId("test1"));
+        for (int i = 0; i < OPERATION_COUNT; i++) {
+            map.containsKey(i);
+        }
 
-        boolean inFirstInstance = address.equals(getNode(instances[0]).getThisAddress());
-        multiMap0.get("test0");
-
-        multiMap0.put("test1", 1);
-        multiMap1.get("test1");
-
-        assertEquals(inFirstInstance ? 1 : 0, multiMap0.getLocalMultiMapStats().getHits());
-        assertEquals(inFirstInstance ? 0 : 1, multiMap1.getLocalMultiMapStats().getHits());
-
-        multiMap0.get("test1");
-        multiMap1.get("test1");
-        assertEquals(inFirstInstance ? 0 : 3, multiMap1.getLocalMultiMapStats().getHits());
+        LocalMapStats stats = geMapStats();
+        assertEquals(OPERATION_COUNT, stats.getOtherOperationCount());
     }
 
     @Test
-    public void testPutStats_afterPutAll() {
-        TestHazelcastInstanceFactory factory = createHazelcastInstanceFactory(2);
-        final HazelcastInstance[] instances = factory.newInstances(getConfig());
-        Map<Integer, Integer> map = new HashMap<Integer, Integer>();
-        for (int i = 1; i <= 5000; i++) {
-            map.put(i, i);
+    public void testOtherOperationCount_entrySet() {
+        Map map = getMap();
+
+        for (int i = 0; i < OPERATION_COUNT; i++) {
+            map.entrySet();
         }
 
-        IMap<Integer, Integer> iMap = instances[0].getMap("example");
-        iMap.putAll(map);
-        final LocalMapStats localMapStats = iMap.getLocalMapStats();
-        assertTrueEventually(new AssertTask() {
-            @Override
-            public void run() throws Exception {
-                assertEquals(5000, localMapStats.getPutOperationCount());
-            }
-        });
+        LocalMapStats stats = geMapStats();
+        assertEquals(OPERATION_COUNT, stats.getOtherOperationCount());
     }
 
     @Test
-    public void testLocalMapStats_withMemberGroups() throws Exception {
-        final String mapName = randomMapName();
-        final String[] firstMemberGroup = {"127.0.0.1", "127.0.0.2"};
-        final String[] secondMemberGroup = {"127.0.0.3"};
+    public void testOtherOperationCount_keySet() {
+        Map map = getMap();
 
-        final Config config = createConfig(mapName, firstMemberGroup, secondMemberGroup);
-        final String[] addressArray = concatenateArrays(firstMemberGroup, secondMemberGroup);
+        for (int i = 0; i < OPERATION_COUNT; i++) {
+            map.keySet();
+        }
 
-        final TestHazelcastInstanceFactory factory = createHazelcastInstanceFactory(addressArray);
-        final HazelcastInstance node1 = factory.newHazelcastInstance(config);
-        final HazelcastInstance node2 = factory.newHazelcastInstance(config);
-        final HazelcastInstance node3 = factory.newHazelcastInstance(config);
-
-        final IMap<Object, Object> test = node3.getMap(mapName);
-        test.put(1, 1);
-
-        assertBackupEntryCount(1, mapName, factory.getAllHazelcastInstances());
+        LocalMapStats stats = geMapStats();
+        assertEquals(OPERATION_COUNT, stats.getOtherOperationCount());
     }
 
     @Test
-    public void testLocalMapStats_preservedAfterEviction() {
-        String mapName = randomMapName();
-        Config config = new Config();
-        config.getProperties().setProperty(GroupProperty.PARTITION_COUNT.getName(), "5");
-        MapConfig mapConfig = config.getMapConfig(mapName);
-        mapConfig.setEvictionPolicy(EvictionPolicy.LRU);
-        MaxSizeConfig maxSizeConfig = mapConfig.getMaxSizeConfig();
-        maxSizeConfig.setMaxSizePolicy(MaxSizeConfig.MaxSizePolicy.PER_PARTITION);
-        maxSizeConfig.setSize(25);
+    public void testOtherOperationCount_values() {
+        Map map = getMap();
 
-        HazelcastInstance instance = createHazelcastInstance(config);
-        IMap<Object, Object> map = instance.getMap(mapName);
-        final CountDownLatch entryEvictedLatch = new CountDownLatch(700);
-        map.addEntryListener(new EntryEvictedListener() {
-            @Override
-            public void entryEvicted(EntryEvent event) {
-                entryEvictedLatch.countDown();
-            }
-        }, true);
-        for (int i = 0; i < 1000; i++) {
-            map.put(i, i);
-            assertEquals(i, map.get(i));
+        for (int i = 0; i < OPERATION_COUNT; i++) {
+            map.values();
         }
-        LocalMapStats localMapStats = map.getLocalMapStats();
-        assertEquals(1000, localMapStats.getHits());
-        assertEquals(1000, localMapStats.getPutOperationCount());
-        assertEquals(1000, localMapStats.getGetOperationCount());
-        assertOpenEventually(entryEvictedLatch);
-        localMapStats = map.getLocalMapStats();
-        assertEquals(1000, localMapStats.getHits());
-        assertEquals(1000, localMapStats.getPutOperationCount());
-        assertEquals(1000, localMapStats.getGetOperationCount());
+
+        LocalMapStats stats = geMapStats();
+        assertEquals(OPERATION_COUNT, stats.getOtherOperationCount());
     }
 
-    private void assertBackupEntryCount(final long expectedBackupEntryCount, final String mapName,
-                                        final Collection<HazelcastInstance> nodes) {
-        assertTrueEventually(new AssertTask() {
-            @Override
-            public void run() throws Exception {
-                long backup = 0;
-                for (HazelcastInstance node : nodes) {
-                    final IMap<Object, Object> map = node.getMap(mapName);
-                    backup += getBackupEntryCount(map);
-                }
-                assertEquals(expectedBackupEntryCount, backup);
-            }
-        });
-    }
+    @Test
+    public void testOtherOperationCount_clear() {
+        Map map = getMap();
 
-    private long getBackupEntryCount(IMap<Object, Object> map) {
-        final LocalMapStats localMapStats = map.getLocalMapStats();
-        return localMapStats.getBackupEntryCount();
-    }
-
-    private String[] concatenateArrays(String[]... arrays) {
-        int len = 0;
-        for (String[] array : arrays) {
-            len += array.length;
+        for (int i = 0; i < OPERATION_COUNT; i++) {
+            map.clear();
         }
-        String[] result = new String[len];
-        int destPos = 0;
-        for (String[] array : arrays) {
-            System.arraycopy(array, 0, result, destPos, array.length);
-            destPos += array.length;
+
+        LocalMapStats stats = geMapStats();
+        assertEquals(OPERATION_COUNT * DEFAULT_PARTITION_COUNT, stats.getOtherOperationCount());
+    }
+
+    @Test
+    public void testOtherOperationCount_containsValue() {
+        Map map = getMap();
+
+        for (int i = 0; i < OPERATION_COUNT; i++) {
+            map.containsValue(1);
         }
-        return result;
+
+        LocalMapStats stats = geMapStats();
+        assertEquals(OPERATION_COUNT * DEFAULT_PARTITION_COUNT, stats.getOtherOperationCount());
     }
 
-    private Config createConfig(String mapName, String[] firstGroup, String[] secondGroup) {
-        final MemberGroupConfig firstGroupConfig = createGroupConfig(firstGroup);
-        final MemberGroupConfig secondGroupConfig = createGroupConfig(secondGroup);
+    @Test
+    public void testOtherOperationCount_isEmpty() {
+        Map map = getMap();
 
-        Config config = getConfig();
-        config.getPartitionGroupConfig().setEnabled(true)
-                .setGroupType(PartitionGroupConfig.MemberGroupType.CUSTOM);
-        config.getPartitionGroupConfig().addMemberGroupConfig(firstGroupConfig);
-        config.getPartitionGroupConfig().addMemberGroupConfig(secondGroupConfig);
-        config.getNetworkConfig().getInterfaces().addInterface("127.0.0.*");
-
-        config.getMapConfig(mapName).setBackupCount(2);
-
-        return config;
-    }
-
-    private MemberGroupConfig createGroupConfig(String[] addressArray) {
-        final MemberGroupConfig memberGroupConfig = new MemberGroupConfig();
-        for (String address : addressArray) {
-            memberGroupConfig.addInterface(address);
+        for (int i = 0; i < OPERATION_COUNT; i++) {
+            map.isEmpty();
         }
-        return memberGroupConfig;
+
+        LocalMapStats stats = geMapStats();
+        assertEquals(OPERATION_COUNT * DEFAULT_PARTITION_COUNT, stats.getOtherOperationCount());
     }
 
-    private <K, V> IMap<K, V> getMap() {
-        HazelcastInstance instance = createHazelcastInstance(getConfig());
-        return instance.getMap(randomString());
+    @Test
+    public void testOtherOperationCount_size() {
+        Map map = getMap();
+
+        for (int i = 0; i < OPERATION_COUNT; i++) {
+            map.size();
+        }
+
+        LocalMapStats stats = geMapStats();
+        assertEquals(OPERATION_COUNT * DEFAULT_PARTITION_COUNT, stats.getOtherOperationCount());
     }
 }
